views changed to work with keras
