+++ ITRI_671/keras_model/keras_metadata.pb
