+++ ITRI_671/keras_model/variables/variables.data-00000-of-00001
