+++ ITRI_671/keras_model/variables/variables.index
